Narrow vertex attributes end-to-end: host types, loader, GL native, Vulkan packed
Host (BufferAttribute):
- AttributeType enum + type-erased type()/data()/byteLength()/cloneUntyped()
- Uint8/Int8/Uint16/Int16BufferAttribute aliases; denormalize() per UNORM/SNORM
- FloatAttributeView: zero-copy float view, widens narrow, de-strides interleaved
- THREEPP_ASSERT debug bounds checks on the element accessors (Assert.hpp)
- compressAttributes(): normal->snorm16, uv->unorm16 ([0,1] only), color->unorm8
- compressSceneAttributes(): scene-wide, skips skinned/displaced/grass/morph/tet
- mergeBufferGeometries merges all scalar types; mergeVertices widens narrow

Loader:
- GLTFLoader keeps COLOR_0 in source width (normalized u8/u16); opt-out flag

GL renderer:
- GLAttributes uploads any scalar type (kills 3 TODO throws); partial updates
  point glBufferSubData at the array instead of copying a sub-vector
- vertexAlphas checks no longer null-deref on narrow color

Vulkan renderer:
- Tier A: attribute reads via FloatAttributeView / untyped version polling;
  narrow host attributes render identically (deforming paths stay float)
- Tier B: static BLAS records pack device buffers - normals oct-snorm16x2
  (12->4 B/vtx), uv unorm16x2, color unorm8x4; decode at the vertex-pull and
  ray-hit fetch sites. Bistro: -114.7 MiB live VRAM, goldens regenerated
  (glass rim pixels moved by oct quantization; PSNR 60.7 dB at the switch)
- GeometryDesc._pad was the sticky moved flag: now flags bit0, packed mask in
  bits 1..3; the per-frame stamp preserves the high bits
- THREEPP_NO_PACK=1 escape hatch; THREEPP_VK_MEMDUMP also dumps at scene-built

Fixes found along the way:
- SkinnedMesh::boneTransform read skinIndex as <int> -> null deref
- Mesh raycast interpolated UVs read vertex B from index c

Demos: bistro/drive/norway_drive/gltf_samples compress static scenery on load.

+++ src/threepp/objects/SkinnedMesh.cpp
@@ -87,7 +87,12 @@ void SkinnedMesh::updateMatrixWorld(bool force) {
 
 void SkinnedMesh::boneTransform(size_t index, Vector3& target) {
 
-    geometry_->getAttribute<int>("skinIndex")->setFromBufferAttribute(_skinIndex, index);
+    // skinIndex is a float attribute everywhere it is produced (GLTFLoader's
+    // readJointIndicesAsFloat, ColladaLoader, and the bones example all create a
+    // FloatBufferAttribute), and the GL/Vulkan skinning paths both read it as
+    // vec4<float>. Asking for <int> made the dynamic_cast fail and this line
+    // dereference a null attribute.
+    geometry_->getAttribute<float>("skinIndex")->setFromBufferAttribute(_skinIndex, index);
     geometry_->getAttribute<float>("skinWeight")->setFromBufferAttribute(_skinWeight, index);
 
     geometry_->getAttribute<float>("position")->setFromBufferAttribute(_basePosition, index);
